+++ src/class.mjs
@@ -1,2 +1,3 @@
 export { default as get_class } from './class/get_class.mjs';
 export { default as is_a } from './class/is_a.mjs';
+export { default as get_parent_class } from './class/get_parent_class.mjs';