+++ tests/class.global.test.mjs
@@ -4,4 +4,5 @@ import '../src/class.global.js';
 test('class.global', () => {
     expect(get_class).not.toBeUndefined();
     expect(is_a).not.toBeUndefined();
+    expect(get_parent_class).not.toBeUndefined();
 });